lrdp_v2: Add function to set gdbmi aslr offset.
Multiarch supporting gdb binary is based on gdb version 10+
has aslr enabled by default. Gdb has an option that should allow
to disable randomization but in this msys version it doesn’t allow that.
Extract head.text entry and subtract the _text address and set it as gdb aslr offset
which need to be added for every return address from gdb.

1. for most targets, this diff will come 0 as gdb somehow takes care of
   this randomization internally and so head.text entry and _text address is same.

2. For targets where its failing (with mutliarch gdb), we will be taking care of this
   by adding gdbmi aslr offset which will come non zero and so proper address will be
   return.

Change-Id: I3c7dad3ce35d81c765d19cf14fb60826b4193763
Signed-off-by: avijayku <quic_avijayku@quicinc.com>