lrdp: Fix iommu page table parsing
Patch adds support to indentify arm smmu version
based on board config. This helps in chosing the
right offsets for various structures based on the
correct header file as struct names are similar
for arm smmu v3 and non v3. GDB is not able to
distinguish to pick a correct header file otherwise.

Change-Id: I5ee0e8d166f2a5e9dbcddb66d5b99afd9f1230a9