lrdp_v2: Resolve compatibility issue for autogvm minidump parse
Since virtio minidump is applied to kernel-5.15. Therefore,
vmid will not be added as a prefix to the name of the region in ELF
by virtio minidump frontend driver. It will cause issue when generate
ELF since the name is different with the actual register name in backend.
To solve this problem, using the prefix name as a judgement condition
to match the right dump files when generate ELF file.

Change-Id: I409b0ed3ea0129856a850503aa1057c92407b696