lrdp_v2: watchdog: Show the time of last pet correctly
If the dog is petted in 1s and barked later due to some issues,
lrdp will show the last pet as below:
Watchdog last pet: 394566250.000

It's confusing. So fix the code to show as below:
Watchdog last pet: 0.39456625

Also, add info to show if the dog is enabled or not.

Change-Id: I80b5b5d8c863db33d64166fc14c5b13f5c7b3ed9