parsers: zram: Add support to override the zram decompression parser
Add the means to let a user give their own decompression program for
LRDP. The program will be forked as a subprocess, must take in
compressed input through stdin and output decompressed output through
stdout (e.g. see `man gzip`). We use --zram_parser_override <path> to
specify the tool of choice.

Note that Windows does not natively support byte-transfer protocols,
so we assume that any decompression binary on Windows accepts the
hexidecimal representation of the binary as a UTF-8 string, and
outputs the decompressed output in the same format.

Change-Id: Ia03b44b90aa2490348f7ba9368c7aef19e8aadc3