lrdp_v2: Remove unnecessary check for hyp elf
When adding --host_vmlinux and --host_mod_path for the SVM parser,
lrdp reads TTBR or VTTBR from corevcpu*.cmm files parsed from PVM.
In the current implementation, an error is reported if --hyp is not
provided, even though it's not required for this SVM parsing method.
Since corevcpu*.cmm provides TTBR or VTTBR, the --hyp check is redundant.
This patch removes the check to allow parsing without requiring --hyp.

Change-Id: I757de759c3f1dccf0098da985a8adfdad90bde32