lrdp_v2: Skip invalid module during traversal of module list
In certain customer dumps, an invalid module may appear in the
module list. While traversing the module list to set up the module
layout, this invalid entry can be added to self.module_layout_dict.
When the DebugImage parser runs, it attempts to unwind the PC for
the secure context, which will trigger an exception and abort DebugImage
parser. This change addresses the issue by preventing the invalid
module.

Change-Id: Icec2c4cfafa888b0e212dca4d08184cc249a26fb